ml-models/requirements.txt
# Python dependencies for ML model
# Install with: pip install -r requirements.txt

numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
python-dotenv==1.0.0
pymongo==4.5.0
